requirements.txt
alembic==1.13.2
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
decorator==5.1.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
executing==2.0.1
filelock==3.15.4
greenlet==3.0.3
gunicorn==23.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
shortcuts==0.11.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
wcwidth==0.2.13
